Fixed reformatting of obsolete entries in catalogs.
When saving a catalog with obsolete (#~ ...) entries, empty line
separating obsolete entries was removed if there weren't any flags or
other metadata associated with the entry.

This patch is the smallest possible change to fix it. "Proper" fix would
be to not touch these parts of the catalog on save at all (related to
reformatting of the entire catalog).

Fixes #329.