Fix list selection visibility on Windows 7.
In Windows 7, selection color in listview that doesn't have focus is
significantly lighter than in previous Windows versions and approaches
the 95%-of-white value that Poedit uses for list shading. This makes it
hard to see where the selection is.

This patch changes shading to use blue color and lighter shade of it.
It's not the best fix (that would be to use focused-selection color and
avoid the unfocusing hack), but it's simple enough to be safe for the
stable series.

Fixes #336.