Fix TM database corruption issues.
Berkeley DB use in Poedit didn't use transaction and so the databases
weren't protected against corruption on program or system crash.
Rewritten BDB code to use DbEnv, which provides transactions and
logging.

It's still possible to corrupt the TM by running multiple instances of
Poedit and touching the TM in parallel. This will be taken care of by
ensuring that only one Poedit instance can run at a time.

Fixes #337 & #352.