Fix menu items capitalization on OS X and GNOME.
On these two platforms, menu items use title-style capitalization for
menu items in English. This is the same as what was used by Windows XP,
but Vista and Windows 7 introduce guidelines that say to use
sentence-style capitalization.

Poedit used sentence-style capitalization until now. This patch fixes it
to use whatever is appropriate on given platform. For Windows, we don't
have different look for XP and Win7; instead, just go for the Win7
style everywhere.