Ignore remembered positions out of screens.
If any window's position as remembered from last time the window was
used is either outside of all connected screens or doesn't fit on its
designated screen, move or resize the window to fit.

This situation can happen e.g. when connecting a notebook to a monitor
and later launching Poedit without one.

Fixes #318.