Build VC++ binaries of Berkeley DB with _BIND_TO_CURRENT_VCLIBS_VERSION.
With VC++ 2008, the default mode of operation is to bind to
Microsoft.VC90.CRT version shipped with the original VC++ 2008 release,
even if the binaries were compiled with SP1. This breaks in certain
situations when using side-by-side deployment (as Poedit does).

This change ensures that the latest CRT is used and everything works as
expected.

See http://msdn.microsoft.com/en-us/library/cc664727.aspx for more.