Use standard Language header instead of X-Poedit-*.
Instead of setting the language and country as separate fields (and
using English names instead of ISO codes at that!), use the standard
Language header introduced in gettext 0.18.

Remove separate fields in catalog properties as well. This looses
auto-completion, so is suboptimal, but we'll eventually replace it with
a better UI that uses iso-codes.