Show errors in-line.
Validate the catalog with msgfmt when saving or when the user explicitly
clicks Validate. Parse msgfmt error output and assign errors to catalog
items. Highlight items with errors in read and show the error in a red
bar under the translation when selected.

Ideally, we'd perform such validations in real-time, but invoking msgfmt
so frequently would be too expensive.