Build gettext binaries as static on OS X.
It is simpler, because we don't have to deal with dylibs, install_name_tool etc. On the other hand, the cost is some 4MB of binary size.

Unfortunately, previous dynamic builds had some runtime problems on 10.5 Leopard (undefined symbol: _error); but, interestingly, not on 10.6+.

Let's revisit this decision and revert-and-improve this commit later, possibly for Poedit 1.6, when we'll require at least 10.6 anyway.