Don't complain about malformed headers when updating from sources.
If the headers differ in any way (e.g. some have Plural-Forms added by
xgettext and some not), msgcat will include all versions, delimited with
"#-#-#-#-# (filename) #-#-#-#-#". Poedit's parser then reports these as
"Malformed header".

To add insult to the injury, Poedit discards the header when it's done
with loading the resulting big catalog, so the error is completely
unnecessary. So just ignore the header when loading this particular file
- it was generated under our control, so it should be safe enough.

Fixes #415.