Handle app-global menu commands in PoeditApp.
Previously, everything was handled at the PoeditFrame level. This didn't
work well on OS X, where the menu is active even with no windows open.

In fact, Poedit crashed because of this. Set app-global menu on OS X to
avoid this.

Also change the behavior when opening files: in GNOME and OS X, open a
new window, as is the standard behavior. (Unless the active window is
empty, in which case reuse it.)

Fixes #474.