Revert "don't compile catalogs into .mo files when saving by default".
This change was causing problems to people not familiar with Poedit, who
followed one of the many tutorials that don't mention this setting, yet
assume the creation of MO files. See e.g.
https://groups.google.com/d/topic/poedit/Coza58B1jQQ/discussion

This change will be re-introduced later, probably for 1.6, but together
with better UI to avoid this kind of confusion. At the very least, there
needs to be explicit "Compile to MO" menu entry before the default can
be changed.

This reverts d0648c95d6a45b37950f1e388a00b8519f806796 commit.