Don't add large dirs to the default TM import set.
Don't add $HOME (or even C:\ on Windows!) to the list of directories to
search when creating a new translation memory.  Scanning the directories
is slow and due to bad UI design, Poedit seems unresponsive.

This is not worth fixing properly, because all this TM import will be
reworked completely.

See #510.