Cleanup deps build on Windows.
Use the same organization as on OS X for gettext builds.

Build as much as possible (wx, bdb, winsparkle) using native projects.

Use VS2012 projects for wx, converted from upstream ones, put into a
different directory.  This is going to be a PITA, but I need convenient
rebuilds of wx.