Create text editing controls in reading order.
Fixes #546.

Quoting the bug:

On Windows, the name reported to assistive technology applications such as screen readers is incorrect for the Source text, Translation and Comment fields. This can be fixed in one of two ways:
1. Ensure that the StaticText for the label is always created immediately before the control it is labelling. This is why Notes for translators gets labelled correctly.
2. If this isn't possible for some reason, use wxAccessible to override the name for these fields. This is somewhat more complicated.
You can test this using the Inspect tool which comes with the Windows SDK.