Don't build GNU gettext on Windows from sources.
Fuck this shit.  Building gettext from sources is notoriously fragile on
anything but Linux and especially so in the MinGW environment.  Every
little change may break the build with spurious or downright mysterious
errors.  Heck, it occasionally breaks *without* any changes to the
environment!  Not only does the build takes hours (let's be fair, this
one is MSYS/Cygwin's fault), it keeps being rebuilt all the time.  Even
after HOURS spent tweaking the build to ignore source(!) files that the
build regularly touches or even modifies(!!), I still can't get it to
behave reliably and sanely.  (And don't get me started on building from
git, that's an order of magnitude worse ordeal thanks to the masochistic
contraption called gnulib.  I gave up on that early on.)

No.  In the interest of preserving the remains of my sanity, I will just
pull a prebuilt nuget package on Windows.  That's not as bad as it
sounds, I maintain it myself at
https://github.com/vslavik/gettext-tools-windows, so I have some control
over it (and building gettext semi-manually is much simpler than doing
it as part of fully automated builds, I tell you).

Unfortunately, this means that the Windows version looses one recently
implemented improvement: No gettext translations on Windows.  This
should be simple in theory, because gettext has relocatable build on
Windows and should be able to find the catalogs in non-fixed location -
if it worked.  I probably wouldn't be able to make it encode the output
in some sensible charset (read: UTF-8) anyway.

So there.