Don't use Boost as submodule.
Will be replaced with subtree-merged copy instead.

One would assume that "boost-release" repository hosted by the official
boostpro GitHub account will be cared for.  Alas, that's not the case,
the content is not kept updated (and when somebody else does the
updates, the pull request just sits there for 7 months and counting,
being ignored).