Display item IDs instead of line numbers.
Instead of completely removing the confusing line numbers, as was the
plan for 1.5, replace them with unique IDs.  This avoids the confusion
of mismatching lines, yet still provides a way to identify individual
entries.  It also opens the way to using symbolic IDs (as is sometimes
done even in gettext).

The plan is to eventually move this to the first column and always use,
instead of being configurable, but that's more work due to various
problems in wxListCtrl.  So this will suffice for now and better UI will
follow after the migration to wxDataViewCtrl.