Don't require immediate saving when creating a file from POT.
Change the workflow to be more in line in what is usually done in
document-based apps: don't immediately save the file and let the user do
it.  To accomplish this, introduce the concept of a file with assigned
filename, but not existing on the disk yet.  If that happens (such as
after creating a file from a POT), ask for filename confirmation when
saving.

Choose appropriate file for the new translation: use lang code for the
name and save directory as the POT file.

Don't do this for completely new catalogs (from sources) yet, because
the whole thing is more complicated there.