Less overwhelming new file workflow.
Previously, File->New was the workflow for setting up scan from sources
and it involved, *immediately* doing the following: 1.  save the file,
2.  setup al properties, 3.  setup sources scanning.  Without any
explanation or any reference material.

Replace this with - hopefully - something friendlier: create a new,
unsaved, empty document and explain why it's empty and what it means.
Offer a link to the manual and explain the two typical actions for
filling in the entries: update from POT or from sources.  Make sure
sources extraction requires file saving (for proper reference to the
path).