Fix multiple Find windows.
Fix the code for maintaining single Find window.  Rather that look it up
by name, use proper code for keeping track of the singleton.  The former
was broken by something, presumably some change in wx, and rather than
hunting the exact cause down, it's faster to do it properly (as it
should have been done in the first place).