OS X Mavericks build fixes.
Mavericks changed the way codesign tool works, it now requires
sub-frameworks and binaries to be signed too. This means signing must be
done as part of build process and cannot be done after the fact.

Still compile Poedit w/o signing by default, but add code to behave
correctly if a signing identity is set.