Fix opening already-open files on case-insensitive systems.
When the user opens a file, Poedit checks if it is already open and
doesn't open another window if the file already has one.  Unfortunately,
this test was broken: it used string comparison, which was case
sensitive (wrong on Windows and OS X) and didn't account for
normalization, symlinks etc.

Use wxFileName::operator== instead.