Improve TM-related messaging in the UI.
Don't say "auto translate", say that it fills from the TM.  Instead of
"Automatic translations", call the suggestions "Suggestions".  Explain
the functionality in preferences so that it is clear it only works with
previously done translations.  Upon translation from TM, show status
message that explains what happened: either than N translations were
filled in and should be reviewed, or that nothing was done and why.