OS X: implement native Open Recent menu.
This code is one big gross ugly hack, but well, it works.

The trouble is that Cocoa does this using a bit of a black magic and
does it immediately after creating the app.  The magic involves finding
the specially-marked Open Recent menu at startup and using it.
Unfortunately, wx apps change the menu quite often (due to it being
per-window, not global), so we have to keep track of this special menu
and reparent it into the currently active menu bar.