OS X: never modify text when setting it into translation text control.
wxTextCtrl implementation on OS X uses insertText:, which is intended
for user input and performs some user input processing, such as
autocorrections.  We need to avoid this, because Poedit's text control
is filled with data when moving in the list control.  Solve this by
using a customized control with overriden DoSetValue().

Fixes #81.