Different indicator of modified entries.
Don't use a star icon, it is confusing to use iconography that has an
established different meaning (favorite).

Instead, use a green line alongside the entries to indicate modified
content, similarly to how e.g.  Visual Studio does it.  A solid line
would be better, but harder to implement with wxListCtrl, so use an icon
that forms reasonably well looking dashed line.