Fix parsing of comments on obsolete entries.
If an obsolete (~-prefixed) entry was preceded by a comment, Poedit's
parser would treat all the following #~ msgid entries as part of the
comment, instead of recognizing them as separate entries.  Because this
super-large comment block would by necessity be the last one in the file
and so wouldn't be associated with an entry, it would be lost when the
file was saved.

Fixes #96.