Make Windows 8 spellchecking configurable.
Behave consistently with the other two platforms in that it is now
possible to disable spellchecking in preferences.  Not doing so was an
omission: USE_SPELLCHECKING wasn't set for Windows builds.

Reorganize the code a bit in the process and move spellchecking-related
helper functions to new spellchecking.{h,cpp} files.