Clear m_textTransPlural in DestroyContentView().
The rest of the code assumes that m_textTransPlural doesn't contain
nullptrs, yet DestroyContentView() kept the array and reset the pointers
in it.  This used to be fine when this function was only called from
PoeditFrame dtor, but no more.

Previously, this sequence of events could cause crash on Windows: open a
file, Ctrl+N, reopen the file.  Now m_textTransPlural would contain
twice as many entries, the first half of them nullptr.