Win: Fix wrong size of textctrl w/ custom font.
If the font was customized, the control's size would sometimes be wrong
- something would set too big minimal size for the bottom translation
text control.  This was apparently related to overzealous blocking of
all wxEVT_TEXT events.  Instead use the same approach used in other
code, disable just the Poedit-specific event handler responsible for the
problem this code was trying to work around in the first place
(disappearing first translation).