Align comments in sidebar with text controls
Align the comments, notes for translators and such so that they are
vertically at the same position as the source text on the left side.  If
there is room, that is: overflow above it if needed.

This makes it easier to read the comments because they are visually
closer to the area read by translators.