TM: return quality score with results
Instead of returning just the match strings, return a struct with
additional information.  Currently, only the quality score is included.

The score is approximated from Lucene's score, with scaling factor to
degrate it if a worse-quality match method is used.  Also ensure that
the best score of 1.0 is only ever returned when the match is exact,
i.e.  the source text is identical in both cases.