Lucene: use non-MT merge scheduler
Attempt to fix the problem with mysteriously corrupted indexes by using
a simpler, single-threaded merge scheduler.  Hopefully it shouldn't
matter for Poedit's use, because its index is relatively small.

The issue seems to be relevant to older Lucene versions (such as 3.0
from Lucene++) and plagues only Windows.  This fix is the same fix
applied by NetBeans:

https://netbeans.org/bugzilla/show_bug.cgi?id=197515