Fix (rare) too many linebreaks in AutoWrappingText
The algorithm assumed that the full width of the control is used for
text layout; that's apparently not the case. Add 4px of buffer to avoid
the bug where the algorithm would sometimes break the text in such a way
that only a single word would be on a line (in the middle of the text).