Fix out of date prefs on OS X when reopened
Because the preferences window is kept in existence once created and
data aren't transfered to it anew, the settings could get out of date
when showing the Preferences for the second (and more) time, if
something in the app changed them in the meantime.

Work around this by reloading them every time the prefs window is
activated or its tab changed. This still isn't perfect and may miss some
changes, but should be good enough for most cases.