Add support for msgmerge --no-fuzzy-matching
msgmerge creates fuzzy entries by default when it sees similar entries
in the "old" version of the file. This is helpful when there's only a
small modification to a string, but on the other hand, it can cause
problems with less experienced translators who accept fuzzy entries
without really checking.

Use --no-fuzzy-matching flag to disable this (and have better
performance), optionally. Bind the behavior to "Consult TM when updating
from sources". It is similar enough to TM updates to make sense, and it
avoids adding yet another setting (to be revisited if there are
complains about this).