Fix setting fonts on OS X
Clicking the checkbox would cause saving preferences _before_ the face
was sent, in which case SaveValues() saved it as unchecked (because the
font was invalid). Then when the user clicked to change the face, a
modal window shows up and when dismissed, the generated activate event
would cause _loading_ of preferences, where the checkbox was saved as
off. This was "only" a problem when setting custom font initially, i.e.
in the common case.

Fix by always having some valid font in there.