Better error when updating w/o source code
A common reason for failure of Update from sources is that there's no
source code - the translator clicks the button without realizing what
it's for. Instead of returning a series of increasingly scary errors,
detect this case and provide a hopefully better understandable
explanation that the source code is missing.