Make "Untranslated entries first" off by default
The setting keeps confusing new users, because the entries are moved
around when you save a file. Ideally, they'd be left in the original
order until the next time the file is opened (or, alternatively,
reordered immediately), but these would be more complicated to implement.

So just change the default to be the unsurprising value instead. The
original behavior is still one click away in the View menu.