Windows: fix undo broken by syntax higlighting.
Syntax hilighting pretty much completely broke Undo/Redo, because the
highlighting changes were considered part of the changes to record. To
fix this, suspend Undo recording while applying style changes.

This is still suboptimal, because it causes undo to work on individual
characters, rather than continuous blocks of text, with no easy
workaround. But it's better than the previous state.