Make undo work with suggestions etc.
Reorganize the code to distinguish between large edits that change the
entire text control via some edit operation such as selecting a
suggestion, copying or clearing the transaction, and between SetValue()
calls that change the displayed value programmatically when changing the
displayed item.

In the former case, the operation should be recorded in the undo stack.
In the latter, it should reset the undo stack and start anew. And that's
exactly what the code does now.