Make the comments text in sidebar selectable
The comments and other information shown in the sidebar uses
wxStaticText, making it non-selectable; this was arguably a regression
compared to 1.6's use of read-only text control.

Make the static text selectable where possible (GTK+ and OS X) and at
least have a context menu with a Copy item where static selectable
labels are impossible (Windows).