Don't forget custom font when pasting on Windows
Text operations that replaced the entire content of the text control
(using a suggestion, copying from sources, or pasting text) would lead
to custom font setting for the text control to be reset. This was
because the richedit control keeps the font as *style* too.

Fix this by explicitly including the font in the default style that is
applied to the entire text after every change in HighlightText().