Symlink pt_BR.lproj to pt.lproj on OS X
OS X uses pt for Brazilian Portuguese and using pt_BR for the
localization causes it to be ignored in some cases (there are two
language settings for Portuguese in OS X Settings: Portuguese
(Brazilian), which is "pt" and is near the top, and Portuguese (Brazil)
that is shown later with a warning about incomplete support and is
"pt_BR". In practice, the former is used, and Apple's own apps ship with
pt.lproj and pt_PT.lproj localizations.

Just do the same, no matter how stupid it sounds. Keep using pt_BR in
Xcode and the files, though, and only create a symlink in the bundle's
Resources. This is the same thing that Sparkle.framework does and it
makes lot of sense.