Fix pasting into selection to replace it on GTK+
When using wxGTK build, pasting into translation field when some part of
it (e.g. the entire text) didn't replace the selection, as it should,
but instead cleared the selection and put the pasted text in front of it.

This appears to be some interaction with freezing/thawing the window. As
this isn't needed under GTK+ for flicker elimination in the first place,
just remove the call there.

Fixes #139.