Fix remembering of bottom splitter's position
Since the introduction of sidebar, the code for remembering - and more
importantly, restoring - position of the splitter separating the list
from the editing part of the window was broken: it always forgot the
setting and went back to the default size.

This was because the code did layout of m_splitter first and then
m_sidebarSplitter. But the former is a child of the latter, so the
sequence has to be the other way around so that m_splitter code works
with final window size.