Add View->Entries with Errors First
Split this setting off from Untranslated Entries First, which is no
longer the default since 1.7. It used to put errors first, then
untranslated and fuzzy. While moving translated entries away was
confusing to some, putting errors at the top should be the default
behavior so that resolving issues found with Validate is easier.