Use Lucene more efficiently in TM code
The existing Lucene code had some defficiencies in the way it approached
index updates. It used a global independent reader that would be
reopened on each query. So far so good, but the reader wasn't bound to
IndexWriter (i.e. wasn't realtime) and a new IndexWriter was created
every time a file was saved to commit translations into the index.

This is inefficient, creating IndexWriter is expensive. Do the
recommended thing instead: keep a single writer active at all times.
Don't make the TranslationMemory::Writer class rollback if not committed
and completely decouple committing from inserting.

Also use realtime IndexReader so that updates to the TM are visible to
it immediately, even before commit(). This will be useful for the
upcoming commit.

See also
http://blog.mikemccandless.com/2011/06/lucenes-near-real-time-search-is-f
ast.html