Update translation memory in realtime
Don't update the TM only when saving the edited file, but also
immediately after the user is done with editing an entry. This makes the
edit immediately available for other translations in the same file that
may be similar.

Still do updates when saving so that the entire content of the file,
even the parts edited previously, are included.

Don't commit realtime submissions, only call Commit() when saving the
file. This provides a reasonable compromise between speed (it wouldn't
be a good idea to fsync() after every edit) and durability (must commit
the data at some point).