Use dispatch_async() fro call_on_main_thread()
On OS X, route dispatching to the main thread through dispatch_async(),
instead of using wxTheApp->CallAfter. The reason for doing this is not
efficiency (it's a little bit worse this way), but Xcode debugger
support: this way, the debugger will include backtrace of the enqueing
code in the backtrace, which makes debugging complicated execution paths
considerably easier.