Defer fetching of Crowdin account info in prefs
Don't fetch account data immediately after creating CrowdinLoginPanel.
Instead, do it when the Accounts preferences tab is selected, to avoid
keychain dialogs when the user has no intention of touching this tab.

This is how it behaved on OS X already thanks to deferred creation of
tabs by wxPreferencesEditor, but not on other platforms.