Fix running gettext in non-ASCII dirs on Windows
Running gettext tools (such as msgfmt or msgcat) in a directory with
non-ASCII filenames that couldn't be represented in whatever codepage
Windows CLI uses (Hebrew filenames are a good example) would cause
gettext CLI tools failures and lead to cryptic errors.

Use the 8.3 compatibility filenames, if available, to pass friendlier
filenames to gettext. Make sure temporary files use such CLI-safe names.
Use temporary file for msgcat formatting too and rename the result to
the target filename, instead of writing to it directly. Make it possible
to use multiple temp files for the same output file in TempOutputFileFor
to faciliate it.