Fix \n auto-insertion with Asian input methods
The naive way \n was inserted when pressing return in 86e2a1e4 was
broken with "complicated" text input methods that use return to confirm
composed text before inserting.

Rework the code to never mess with keystrokes. Instead, remember than
return was pressed and if it was _and_ the current edit position
contains a newline not preceeded by literal \n, insert it.