Use Singular/Plural labels for "simple" languages
Don't use descriptive CLDR-style names for plural forms (One, Zero,
Other etc.) when the language has only two plural forms, like English
does. If that's the case, use Singular and Plural labels instead, so
that the source text and translation's labels are the same. There's no
risk of confusion for these languages, so it's better to do it like this
instead of the more accurate, but unnecessary CLDR-style labels.