Fix background_queue deadlock on showdown on Win
ThreadPool being a global object meant that a) it was always created,
even when not used and b) shutdown of worker threads was done as part of
CRT deinitialization, which was apparently problematic on Windows,
resulting in deadlock and never terminating process.

Instead, initialize the pool lazily and explicitly destroy it as part of
PoeditApp shutdown sequence, as is done with other components.