Set locale on all platforms
Previously, the locale was only set under GTK+ where it matters the
most, but do it on all platforms, for correct formatting of dates etc.
Fail gracefully to the old translation-only setup with the C locale if
the locale couldn't be set (e.g. language not supported by the OS -
Windows, I'm looking at you).