Ensure that only one instance of Poedit is running
Make sure that two instances are not running at the same time on Windows
and Unix. Use wxSingleInstanceChecker together with lightweight IPC
mechanism to pass commands (open files, open URI, activate) to the
primary instance if launched for the 2nd+ time and terminate
immediately. This is the standard behavior of GTK+ 3 apps too.

Ideally the GTK+ version would use GtkApplication, which has all this
built-in, but until wx gains support for it, it doesn't seem to be worth
the effort.

On OS X, this is handled automatically by OS and wx.