Better detect source code availability
Improve HasSourcesAvailable() to check for the existence of directories
to be searched. This should improve error reporting and prevent mistaken
updates from sources spanning entire disk.

Also split the check into two: HasSourcesConfigured() is efficient and
checks if source code extraction was setup at all (and if not, the
toolbar button is disabled) while HasSourcesAvailable() is more
expensive and should be called prior to updating from sources (and show
"source code not available" error as appropriate).