Improve UI for setting source extraction paths
Remove the old and cryptic, user unfriendly controls where everything
had to be typed manually, as relative paths, with no indication of the
expected content.

The new UI uses lists where the user can add or drag & drop files and
folders and they will be displayed as correct relative paths
automatically. Basepath is no longer editable, but it is still shown
because wildcards need to be relative to it.