Fix setting of UI language that has no OS locale support
The logic intended for this was broken. Fixed to behave as expected: set
locale for OS default and UI language to the chosen one, even if it has
no corresponding locale and setting the default locale didn't fail.

For example: running OS in Spanish and having UI language set to
Aragonese on Windows.