Fix instability in translation memory queries
While the Lucene classes used are in theory thread-safe, in practice
there's a race condition related to creation of IndexSearcher and
IndexReader instances. Readers use explicit refcounting and it could
happen that a subreader would be closed when used by another thread.

See https://issues.apache.org/jira/browse/LUCENE-3567 for the exact
issue encountered by Poedit's use as well.

Fix this by using a manager that ensures refcounting of in-use readers.
Also switch to using a single shared IndexSearcher, which is the
recommended approach, instead of re-creating it on every use.