Improve source lang detection in mixed lang cases
CLD2, used by Poedit to detect source language, heavily penalizes
English in bilingual content in some cases as "boilerplate" because it
is tailored for the web. So e.g. 66% English, 33% Italian is tagged as
Italian.

Poedit's bias should be the opposite: English is almost always the
correct answer for PO source language. We need to manually undo CLD2's tweaking of the scores to get English on equal footing.