Use real browser for Crowdin auth on Windows & GTK+
There are some hard-to-track crashes with some MSIE versions,
apparently, and it's more secure to use the default browser instead of
embedded one anyway. It is also more cumbersome, though, as the user has
to confirm opening a custom URL. That's one-time annoyance only, though.

Add x-schema-handler/poedit to a new poedit-uri.desktop file to do the
same on Unix.