Fix saving of files without path
When a file is opened with e.g. "poedit pl.po", the filename has no path
component and wxFileName::SplitPath() returns empty string for it. Make
sure it is correctly treated as . instead in all uses.

This was causing failures to save a file, because Poedit would try to
create temporary files in / instead of in the file's directory, which of
course failed on any non-instanely setup system.