FixDuplicatesIfPresent:
It's true that Poedit never produces PO files broken in this way, but it
could *edit* such files without fixing them prior to 1.8, and so
X-Generator could be Poedit. Whitelist only recent versions of Poedit.

Notice that the check doesn't include Poedit 1.4 or older, because these
old versions of Poedit didn't emit the X-Generator header yet.