Fix reformatting of POs to CRLF to work in sandbox
The code to convert PO files back to DOS lineendings after passing them
through msgcat wouldn't work in restricted environments such as OS X
sandbox because of its use of temp files alongside the final output file.

Do the same thing as in other places: do this processing on temporary
files that are then renamed to the final output file. This should have
been done from the beginning, but this particular place was overlooked.