Nudge Windows 7 users towards upgrading
Poedit's spellchecking support on Windows requires version 8 or newer,
because that's when the OS spellchecking API was introduced. Because
experience using Poedit is worse w/o this functionality, it would be
best if as many Poedit users as possible used modern versions of Windows.

And because Microsoft is making the upgrade to Windows 10 free for 7+
users, I can do a little bit to increase the chances of people using
Poedit at its best: tell them that upgrading to Windows 10 will improve
the app.