Don't query suggestions when idly scrolling through file
Apparently, some users are in the habbit of holding down the down arrow
key and scrolling through the file like that. This would cause totally
unnecessary suggestion queries in Poedit on each selection change.

Detect this situation by checking if less than 100ms passed since
previous selection change. If that's the case, delay the query by
up to 100ms after the rapid selection changes finished.

Also fixes ugly flicker in the sidebar in this situation.