Preserve permissions when saving files on UNIX
Don't reset the permissions to their default values as a side effect to
creating a new temporary file and renaming it into the destination. Read
the original file's permissions and ownership, if it exists, and apply
them to the new file.

On OS X, use NSFileManager replaceItemAtURL: that does this (and more)
automatically.