Include relative positions in Lucene PhraseQuery
Don't loose the information about skipped stopwords from the token
stream.

This fixes a problem where even a perfect match wouldn't be found in the
TM in some cases - specifically when the phrase contains stopwords and
there's another stored string that contains the same phrase without the
stopwords. For example, "Abrir a la venta" wouldn't be found previously
if there was another string containing "Abrir la venta" inside it.