Resolved an issue with the handling of the \a (bell) character.
Freedom Scientific has reported a bug with the handling of the \a (bell)
character in PO files. In the PO file used by JAWS, the \a character is
often used as a delimiter. The \a characters are typically replaced with
spaces.

In order to resolve the bug, I modified the
AnyTranslatableTextCtrl::EscapePlainText and
AnyTranslatableTextCtrl::UnescapePlainText functions to properly handle
the \a character.

This bug was originally reported as bug 88272 {POEdit: When inserting
the \a (or bell character) in a string, it gets converted to another
character}.