Use mCtrl menubar for nicer look on Windows 10
Native menus have ugly, unremovable separator between the menu and the
toolbar. The only way to get rid of it is to use own implementation of
menubar as used e.g. by older versions of MSIE or other rebar-using apps.

Fortunately, there's just such an implementation in mCtrl
(http://mctrl.org) and it's relatively easy to integrate it with
wxWidgets.