Provide sane JSON API in http_client
Instead of the homegrown and somewhat crazy json_dict API, use the
nlohmann::json library and return the usual kind of json objects.

The old API had some justification in that it was just a thin wrapper
around native JSON parsing. Give up on that and use the same parser on
all platforms (it's faster than C++REST SDK's one anyway).