Match Explorer's list header appearance on Win10
Particularly on Windows 10 in HiDPI mode, the native header control has
different dimensions from what the user is used to in Explorer. Enlarge
the header slightly to match Explorer and importantly, to visually
separate the header from content in HiDPI mode.

Also remove the slightly enlarged header font that was previously used
for the purpose of separation with wxListCtrl.