Don’t use http_client_config when not needed
Don’t pass default http_client_config to http_client if not needed, i.e.
elsewhere than Windows.

This fixes a crash in C++REST SDK reported by Debian. Cause is unknown,
but related to http_client_config and when that is not used on Linux
anyway, why make things harder for everybody.