Introduce Config class for accessing user options
Add a new header with higher-level abstraction over wxConfig to make
access to configuration options easier. In particular, this API is
thread-safe (unlike wxConfig) and doesn’t require the caller to provide
defaults.

Only use it for UseTM flag for now, but all settings will eventually be
migrated to it.