Show "Pre-translated" tag news to Needs Review toggle
If an item was pre-translated, show explanatory tag need the Needs
Review toggle, so that it is clear what the translation's state is and
why it needs review. Remove the pre-translation icon from the list where
it only clutters display and is redundant (since the entries are already
marked with orange there).