Fix broken selection after using a suggestions (Win)
After pressing Ctrl+1 with focus in the text control, selection wasn't
shown until changing focus elsewhere and back (e.g. by double Alt+Tab).
The cause of this misbehavior of the control is unclear, but there was
no reason to mess with the selection in the first place: just clear the
text instead.