Use dedicated extractor for gettext
Have a dedicated Extractor implementation for languages supported by GNU
gettext tools, instead of using a "legacy" extractor. This allows
greater flexibility in invocation and enables future improvements.

As the first improvement, files to extract from are now passed through a
file to xgettext and not as command-line arguments. This removes the old
limitation to the number of files processed at once due to command line
length limits on some platforms and makes it possibly to execute
xgettext only once, as well as to avoid concatenating files with msgcat.
Consequently, extraction is noticeably faster, particularly on platforms
where launching a child process is expensive (read: Windows).

Also fixes #136.