Don't rescale bitmaps for Windows' 125% zoom
The most common (not-so-)HiDPI mode on Windows is 125% and Poedit's
automatically resized bitmaps look particularly ugly in it, with not
much of a size difference, but distinct fuzziness to them.

This is completely unnecessary, because bitmaps look fine at their
original size despite slightly larger text. Just use bitmaps unmodified
unless the scaling factor is larger.