Invoke DDE operations asynchronously
When reacting to DDE commands on Windows, sent when poedit.exe is
invoked while already running, return as soon as possible and perform
the actual operation at idle time.

The reason for this is that there's a timeout on DDE operations and if
something takes longer to do (e.g. loading a PO file), it may show a
confusing error message. This way we limit synchronous inter-process
communication to the minimum.

See also the discussion in #358.